pyproject.toml
[tool.poetry]
name = "fast_api_template"
version = "0.0.1"
description = "Microservice template on FastAPI"
authors = ["Артур Сагитуллин <asagitullin@sdvor.com>", "Динар Шаргабдуллин <dshargabdullin@sdvor.com>"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3

[[tool.poetry.source]]
name = "nexus"
url = "https://nexus.itlabs.io/repository/pypi-all/simple/"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
#с версии 1.5.8 обновился и не работает с aioredis
sentry-sdk = "1.5.7"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
psycopg2 = "^2.9.1"
asyncpg = "^0.24.0"
asyncio = "^3.4.3"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.5"
structlog = "^21.2.0"
opentelemetry-instrumentation-fastapi = "^0.25b2"
opentelemetry-instrumentation-asyncpg = "^0.25b2"
opentelemetry-sdk = "^1.6.2"
opentelemetry-instrumentation = "^0.25b2"
aioredis = "^2.0.0"
prometheus-fastapi-instrumentator = "^5.7.1"
opentelemetry-exporter-jaeger = "^1.6.2"
tortoise-orm = "^0.17.8"
aerich = "^0.5.8"
fastapi-admin = "^1.0.3"
pylint = "^2.12.1"
cffi = "^1.14.6"
opencv-python = "^4.6.0"
apscheduler = "^3.9.1"
aiobotocore = "^2.3.3"
aiofiles = "^0.8.0"


[tool.poetry.dev-dependencies]
pytest = "6.2.5"
black = "^21.9b0"
isort = "^5.9.3"
httpx = "^0.20.0"
pytest-asyncio = "^0.16.0"
cookiecutter = "^1.7.3"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
